% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animals.R
\name{cats}
\alias{all_animals}
\alias{all_baby_animals}
\alias{arthropods}
\alias{baby_farm_animals}
\alias{baby_rodents}
\alias{birds}
\alias{cats}
\alias{chicks}
\alias{dogs}
\alias{farm_animals}
\alias{foals}
\alias{horses}
\alias{kittens}
\alias{other_animals}
\alias{other_baby_animals}
\alias{puppies}
\alias{reptiles}
\alias{rodents}
\alias{sea_animals}
\title{urls for pictures of cats}
\usage{
cats()

kittens()

dogs()

puppies()

horses()

foals()

rodents()

baby_rodents()

birds()

chicks()

arthropods()

farm_animals()

baby_farm_animals()

reptiles()

sea_animals()

other_animals()

other_baby_animals()

all_animals()

all_baby_animals()
}
\description{
urls for pictures of cats

urls for pictures of kittens

urls for pictures of dogs

urls for pictures of puppies

urls for pictures of horses

urls for pictures of foals

urls for pictures of rodents

urls for pictures of rodents

urls for pictures of birds

urls for pictures of chicks

urls for pictures of arthropods

urls for pictures of farm animals

urls for pictures of farm animals

urls for pictures of reptiles

urls for sea animals pictures

urls for pictures of other animals

urls for other baby animals pictures

urls for all animal pictures

urls for all baby animal pictures
}

