% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{animals_raw}
\alias{animals_raw}
\title{Raw  Animal description dataset}
\format{
A \link[tibble:tibble-package]{tibble} with 610 rows and 48 variables.
}
\source{
\url{https://a-z-animals.com/}
}
\usage{
animals_raw
}
\description{
Raw  Animal description dataset
}
\keyword{datasets}
