% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_dict}
\alias{data_dict}
\title{Animals dataset data dictionary}
\format{
A \link[tibble:tibble-package]{tibble} with 48 rows and 3 variables.
}
\source{
\url{https://a-z-animals.com/}
}
\usage{
data_dict
}
\description{
Animals dataset data dictionary
}
\keyword{datasets}
