% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zoo_tycoon}
\alias{zoo_tycoon}
\title{Animal data from Video Game Zoo Tycoon}
\format{
A \link[tibble:tibble-package]{tibble} with 119 rows and
24 variables.

\describe{
\item{name}{Character, name of animal.}
\item{sci_name}{Character, scientific name of animal. Will be \code{NA} for
mythical creatures.}
\item{biome}{Character, animal's natural habitat.}
\item{favorite_foliage}{Character, animal's favorite foliage. 2 values are
\code{NA}.}
\item{location}{Character, geographical location of origin.}
\item{cost}{Integer, cost in in-game dollars.}
\item{initial_happiness}{Integer, initial happiness values. Takes values
between 10 and 90.}
\item{min_suitability}{Integer, Animals get a happiness penalty if their
exhibit suitability score is lower than this number, }
\item{climbs_fences}{Character, Whether the animal can climb fences, takes
values \code{"No"} and \code{"Yes"}.}
\item{jumps_fences}{Character, Whether the animal can jump fences, takes
values \code{"No"} and \code{"Yes"}.}
\item{diet}{Character, what food the animal eats.}
\item{shelters}{Character, comma seperated names of shelters the animal
likes.}
\item{minimum_social_group}{Integer, minimum social group.}
\item{animal_density}{Numeric, animal density.}
\item{adult_attractiv}{Integer, attractiveness of adult animals, the higher
this number, the more popular the animal is with guests.}
\item{young_attractiv}{Integer, attractiveness of young animals, the higher
this number, the more popular the animal is with guests.}
\item{lifespan}{Numeric, lifespan in months.}
\item{death_chance}{Numeric, death chance, mechanics unknown.}
\item{puberty}{Numeric, puberty length in months.}
\item{reproduction_threshold}{Integer, threshold based on happiness on
whether the animal will reproduce.}
\item{reproduction_chance}{Character, chance of a successful breeding,
takes values \code{"High"}, and \code{"Low"}.}
\item{reproduction_interval}{Numeric, time between reproduction attempts,
in months.}
\item{number_of_offspring}{Integer, number of offspring or eggs produced.}
\item{expansion}{Character, name of expansion where animal is from.}
}
}
\source{
\url{https://zootycoon.fandom.com/wiki/Main_Page}
}
\usage{
zoo_tycoon
}
\description{
Contains all the data for all animals in base game and expansions.
}
\keyword{datasets}
