% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{animals}
\alias{animals}
\title{Cleaned Animal description dataset}
\format{
A \link[tibble:tibble-package]{tibble} with 610 rows and 2 variables.
}
\source{
\url{https://a-z-animals.com/}
}
\usage{
animals
}
\description{
Cleaned Animal description dataset
}
\keyword{datasets}
